+++ omnimatter_fluid/data.lua
@@ -1,1 +1,6 @@
-require("function")
+if not omni then omni = {} end
+if not omni.fluid then omni.fluid={} end
+
+require("functions")
+
+require("prototypes.sluids-boiler")
